README update to show labor division
We did not read the section in our assignment details that said github commits would show participation. Because of this we only committed the code one time. We did meet in person on multiple occasions for many hours to create the final project. This readme now lays out the exact division of labor we used while working on the project